+++ DataModels/Models/Registration.cs
@@ -9,7 +9,7 @@ namespace DataModels.Models
         public int Id { get; set; }
         public DateTime DateOfRegistration { get; set; }
         public string TechnicalCheckService { get; set; }
-        public double Kilometrage { get; set; }
+        public long Kilometrage { get; set; }
         public int TotalPrice { get; set; }
         public string AdditionalComment { get; set; }
